+++ src/onyx/log/commands/seal_output.clj
@@ -23,6 +23,7 @@
             (update-in [:jobs] (fn [coll] (remove (partial = (:job args)) coll)))
             (update-in [:jobs] vec)
             (update-in [:completed-jobs] conj (:job args))
+            (update-in [:completed-jobs] vec)
             (update-in [:allocations] dissoc (:job args))
             (update-in [:peer-state] merge (into {} (map (fn [p] {p :idle}) peers)))
             (reconfigure-cluster-workload)))