+++ src/onyx/static/default_vals.clj
@@ -9,11 +9,13 @@
    ; task defaults
    :onyx/batch-timeout 1000
 
+   ; zookeeper defaults
+   :onyx.peer/zookeeper-timeout 6000
+
    ; peer defaults
    :onyx.peer/inbox-capacity 1000
    :onyx.peer/outbox-capacity 1000
    :onyx.peer/drained-back-off 400
-   :onyx.peer/sequential-back-off 2000
    :onyx.peer/job-not-ready-back-off 500
 
    ; messaging defaults