+++ test/onyx/log/generative_peer_join.clj
@@ -10,6 +10,7 @@
             [clojure.test.check.generators :as gen]
             [clojure.test.check.properties :as prop]
             [clojure.test :refer :all]
+            [com.gfredericks.test.chuck :refer [times]]
             [com.gfredericks.test.chuck.clojure-test :refer [checking]]))
 
 (def onyx-id (java.util.UUID/randomUUID))
@@ -95,7 +96,7 @@
 (deftest greedy-allocation
   (checking
    "Checking greedy allocation causes all peers to be allocated to one of two jobs"
-   50
+   (times 50)
    [{:keys [replica log peer-choices]} 
     (log-gen/apply-entries-gen 
      (gen/return
@@ -103,7 +104,7 @@
                  :messaging {:onyx.messaging/impl :dummy-messenger}}
        :message-id 0
        :entries (assoc
-                    (log-gen/generate-join-entries (log-gen/generate-peer-ids 8))
+                    (log-gen/generate-join-queues (log-gen/generate-peer-ids 8))
                   :job-1 [(api/create-submit-job-entry job-1-id
                                                        peer-config 
                                                        job-1 
@@ -124,14 +125,14 @@
 (deftest greedy-allocation-reallocated
   (checking
    "Checking peers reallocated to other job when killed"
-   50
+   (times 50)
    [{:keys [replica log peer-choices]} 
     (log-gen/apply-entries-gen 
      (gen/return
       {:replica {:job-scheduler :onyx.job-scheduler/greedy
                  :messaging {:onyx.messaging/impl :dummy-messenger}}
        :message-id 0
-       :entries (assoc (log-gen/generate-join-entries (log-gen/generate-peer-ids 8))
+       :entries (assoc (log-gen/generate-join-queues (log-gen/generate-peer-ids 8))
                   :job-1 [(api/create-submit-job-entry job-1-id
                                                        peer-config 
                                                        job-1 
@@ -156,7 +157,7 @@
       {:replica {:job-scheduler :onyx.job-scheduler/balanced
                  :messaging {:onyx.messaging/impl :dummy-messenger}}
        :message-id 0
-       :entries (assoc (log-gen/generate-join-entries (log-gen/generate-peer-ids 6))
+       :entries (assoc (log-gen/generate-join-queues (log-gen/generate-peer-ids 6))
                   :job-1 [(api/create-submit-job-entry job-1-id
                                                        peer-config 
                                                        job-1 
@@ -173,14 +174,14 @@
 (deftest balanced-allocations-uneven
   (checking
    "Checking Balanced allocation causes peers to be evenly over tasks when the spread is uneven"
-   50
+   (times 50)
    [{:keys [replica log peer-choices]}
     (log-gen/apply-entries-gen
      (gen/return
       {:replica {:job-scheduler :onyx.job-scheduler/balanced
                  :messaging {:onyx.messaging/impl :dummy-messenger}}
        :message-id 0
-       :entries (assoc (log-gen/generate-join-entries (log-gen/generate-peer-ids 7))
+       :entries (assoc (log-gen/generate-join-queues (log-gen/generate-peer-ids 7))
                   :job-1 [(api/create-submit-job-entry job-1-id
                                                        peer-config
                                                        job-1
@@ -200,14 +201,14 @@
 (deftest balanced-allocations
   (checking
    "Checking balanced allocation causes peers to be evenly split"
-   50
+   (times 50)
    [{:keys [replica log peer-choices]} 
     (log-gen/apply-entries-gen 
      (gen/return
       {:replica {:job-scheduler :onyx.job-scheduler/balanced
                  :messaging {:onyx.messaging/impl :dummy-messenger}}
        :message-id 0
-       :entries (assoc (log-gen/generate-join-entries (log-gen/generate-peer-ids 12))
+       :entries (assoc (log-gen/generate-join-queues (log-gen/generate-peer-ids 12))
                   :job-1 [(api/create-submit-job-entry job-1-id
                                                        peer-config 
                                                        job-1 
@@ -227,11 +228,152 @@
    (is (= (map count (vals (get (:allocations replica) job-2-id))) [2 2 2]))
    (is (= (map count (vals (get (:allocations replica) job-3-id))) []))))
 
+(deftest job-percentages-balance
+  (checking
+    "Checking percentages allocation causes peers to be evenly split"
+    (times 50)
+    [{:keys [replica log peer-choices]} 
+     (let [percentages-peer-config (assoc peer-config 
+                                          :onyx.peer/job-scheduler 
+                                          :onyx.job-scheduler/percentage)] 
+       (log-gen/apply-entries-gen 
+         (gen/return
+           {:replica {:job-scheduler :onyx.job-scheduler/percentage
+                      :messaging {:onyx.messaging/impl :dummy-messenger}}
+            :message-id 0
+            :entries (assoc (log-gen/generate-join-queues (log-gen/generate-peer-ids 20))
+                            :job-1 [(api/create-submit-job-entry job-1-id
+                                                                 percentages-peer-config 
+                                                                 (assoc job-1 :percentage 30) 
+                                                                 (planning/discover-tasks (:catalog job-1) 
+                                                                                          (:workflow job-1)))]
+                            :job-2 [(api/create-submit-job-entry job-2-id
+                                                                 percentages-peer-config 
+                                                                 (assoc job-2 :percentage 30) 
+                                                                 (planning/discover-tasks (:catalog job-2) 
+                                                                                          (:workflow job-2)))]
+                            :job-3 [(api/create-submit-job-entry job-3-id
+                                                                 percentages-peer-config 
+                                                                 (assoc job-3 :percentage 40) 
+                                                                 (planning/discover-tasks (:catalog job-3) 
+                                                                                          (:workflow job-3)))
+                                    {:fn :kill-job :args {:job job-3-id}}])
+            :log []
+            :peer-choices []})))]
+    (is (= (map count (vals (get (:allocations replica) job-1-id))) [2 2 2]))
+    (is (= (map count (vals (get (:allocations replica) job-2-id))) [2 2 2]))
+    (is (= (map count (vals (get (:allocations replica) job-3-id))) []))))
+
+(def job-1-pct-tasks
+  {:workflow [[:a :b] [:b :c]]
+   :catalog [{:onyx/name :a
+              :onyx/ident :core.async/read-from-chan
+              :onyx/type :input
+              :onyx/percentage 25
+              :onyx/max-peers 1
+              :onyx/medium :core.async
+              :onyx/batch-size 20
+              :onyx/doc "Reads segments from a core.async channel"}
+
+             {:onyx/name :b
+              :onyx/fn :mock/fn
+              :onyx/percentage 37.5
+              :onyx/type :function
+              :onyx/batch-size 20}
+
+             {:onyx/name :c
+              :onyx/ident :core.async/write-to-chan
+              :onyx/type :output
+              :onyx/percentage 37.5
+              :onyx/medium :core.async
+              :onyx/batch-size 20
+              :onyx/doc "Writes segments to a core.async channel"}]
+   :task-scheduler :onyx.task-scheduler/percentage})
+
+(def job-2-pct-tasks
+  {:workflow [[:d :e] [:e :f]]
+   :catalog [{:onyx/name :d
+              :onyx/ident :core.async/read-from-chan
+              :onyx/type :input
+              :onyx/percentage 25
+              :onyx/medium :core.async
+              :onyx/batch-size 20
+              :onyx/doc "Reads segments from a core.async channel"}
+
+             {:onyx/name :e
+              :onyx/fn :mock/fn
+              :onyx/percentage 25
+              :onyx/type :function
+              :onyx/batch-size 20}
+
+             {:onyx/name :f
+              :onyx/ident :core.async/write-to-chan
+              :onyx/type :output
+              :onyx/percentage 50
+              :onyx/medium :core.async
+              :onyx/batch-size 20
+              :onyx/doc "Writes segments to a core.async channel"}]
+   :task-scheduler :onyx.task-scheduler/percentage})
+
+(def job-3-pct-tasks
+  {:workflow [[:g :h] [:h :i]]
+   :catalog [{:onyx/name :g
+              :onyx/ident :core.async/read-from-chan
+              :onyx/type :input
+              :onyx/percentage 25
+              :onyx/medium :core.async
+              :onyx/batch-size 20
+              :onyx/doc "Reads segments from a core.async channel"}
+
+             {:onyx/name :h
+              :onyx/fn :mock/fn
+              :onyx/type :function
+              :onyx/percentage 25
+              :onyx/batch-size 20}
+
+             {:onyx/name :i
+              :onyx/ident :core.async/write-to-chan
+              :onyx/type :output
+              :onyx/percentage 50
+              :onyx/medium :core.async
+              :onyx/batch-size 20
+              :onyx/doc "Writes segments to a core.async channel"}]
+   :task-scheduler :onyx.task-scheduler/percentage})
+
+(deftest percentage-task-allocations
+  (checking
+   "Checking percentage task allocations"
+   (times 50)
+   [{:keys [replica log peer-choices]} 
+    (log-gen/apply-entries-gen 
+     (gen/return
+      {:replica {:job-scheduler :onyx.job-scheduler/balanced
+                 :messaging {:onyx.messaging/impl :dummy-messenger}}
+       :message-id 0
+       :entries (assoc (log-gen/generate-join-queues (log-gen/generate-peer-ids 16))
+                       :job-1 [(api/create-submit-job-entry job-1-id
+                                                            peer-config 
+                                                            job-1-pct-tasks 
+                                                            (planning/discover-tasks (:catalog job-1-pct-tasks) (:workflow job-1-pct-tasks)))]
+                       :job-2 [(api/create-submit-job-entry job-2-id
+                                                            peer-config 
+                                                            job-2-pct-tasks 
+                                                            (planning/discover-tasks (:catalog job-2-pct-tasks) (:workflow job-2-pct-tasks)))]
+                       :job-3 [(api/create-submit-job-entry job-3-id
+                                                            peer-config 
+                                                            job-3-pct-tasks
+                                                            (planning/discover-tasks (:catalog job-3-pct-tasks) (:workflow job-3-pct-tasks)))
+                               {:fn :kill-job :args {:job job-3-id}}])
+       :log []
+       :peer-choices []}))]
+   (is (= (sort (map count (vals (get (:allocations replica) job-1-id)))) [1 3 4]))
+   (is (= (sort (map count (vals (get (:allocations replica) job-2-id)))) [2 2 4]))
+   (is (= (map count (vals (get (:allocations replica) job-3-id))) []))))
 
 (deftest peer-leave
   (checking
     "Checking balanced allocation causes peers to be evenly split"
-    1000
+    (times 50)
     [{:keys [replica log peer-choices]} 
      (log-gen/apply-entries-gen 
        (gen/return
@@ -239,7 +381,7 @@
                     :messaging {:onyx.messaging/impl :dummy-messenger}}
           :message-id 0
           :entries 
-          (-> (log-gen/generate-join-entries (log-gen/generate-peer-ids 4))
+          (-> (log-gen/generate-join-queues (log-gen/generate-peer-ids 4))
               (assoc :leave-anytime [{:fn :leave-cluster 
                                       :args {:id :p1} 
                                       :immediate? true}]))