+++ test/onyx/log/generators.clj
@@ -131,12 +131,15 @@
                   g
                   (apply-entries-gen (apply-entry-gen g)))))))
 
-(defn generate-join-entries [peer-ids]
+(defn build-join-entry [peer-id]
+  {:fn :prepare-join-cluster
+   :immediate? true
+   :args {:peer-site (extensions/peer-site messenger) 
+          :joiner peer-id}})
+
+(defn generate-join-queues [peer-ids]
   (zipmap peer-ids
-          (map (fn [id] [{:fn :prepare-join-cluster
-                         :immediate? true
-                         :args {:peer-site (extensions/peer-site messenger)
-                                :joiner id}}])
+          (map (comp vector build-join-entry) 
                peer-ids)))
 
 (defn generate-peer-ids [n]