+++ changes.md
@@ -20,6 +20,7 @@
 - New feature: Flux policies
 - New feature: `:onyx/min-peers` can be set on grouping tasks to create a lower bound on the number of peers required to start a task
 - New documentation: Event context map information model
+- New documentation: Default configuration & timeout values
 - API: Added metadata to all public API functions indicating which Onyx version they were added in
 - API: `onyx.api/start-peers!` API renamed to `onyx.api/start-peers`
 - API: `onyx.api/shutdown-peers` is now idempotent